finished with the edit and delete highlight

+++ server/endpointsJs/highlights/deleteHighlight.js
@@ -1,6 +1,7 @@
 const Highlight = require("../../schemas/highlightSchema");
 const User = require("../../schemas/userSchema");
 
+
 async function deleteHighlight(req, res) {
     const highlightId = req.params.highlightId;
     const {userId} = req.body;
@@ -13,7 +14,7 @@ async function deleteHighlight(req, res) {
         }
 
     } catch {
-        return res.status(500).json({ error: 'Internal Server Error -> (Searching for the user)' }); 
+        return res.status(500).json({ error: 'Internal Server Error - Searching for the user' }); 
     }
 
     try {
@@ -24,18 +25,18 @@ async function deleteHighlight(req, res) {
         }
 
     } catch {
-        return res.status(500).json({ error: 'Internal Server Error -> (Searching for the highlight)' }); 
+        return res.status(500).json({ error: 'Internal Server Error -  Searching for the highlight' }); 
     }
 
-    if (highlight.ownerId.toString() !== userId.toString()) {
-        return res.status(500).json({ error: 'You are not the creator of this Highlight!' });
+    if (!(highlight.ownerId.toString() === userId.toString() || user.isAdministrator)) {
+        return res.status(400).json({error: "You cannot delete this Highlight!"});
     }
     
     try {
         await Highlight.findOneAndDelete({_id: highlightId});
         
     } catch {
-        return res.status(500).json({ error: 'Internal Server Error -> (Deleting the Highlight)' });
+        return res.status(500).json({ error: 'Internal Server Error - Deleting the Highlight' });
     }
 
     return res.status(200).json({ message: 'Highlight deleted successfully!' });
